Prevalence of Creative Commons licenses by variant, year and journal

This dataset contains the number and proportion of articles with Creative Commons license (CC) by license variant and year for hybrid journals as listed by the German Open Access monitor since 2017.

Usage

cc_jn_ind

Format

An object of class tbl_df (inherits from tbl, data.frame) with 43752 rows and 6 columns.

Details

Journal's article volume was calculated using the Crossref database snapshot. Note that only articles published in regular issues aside from supplements containing conference contributions like meeting abstracts, indicated by non-numeric pagination were included. Also, non-scholarly journal content, such as the table of contents were excluded, following Unpaywall's paratext recognition approach, which was expanded to include patterns indicating corrections.

CC licenses were also identified through Crossref metadata records. License information for author accepted manuscripts ("aam") were not considered.

Variables:

cr_journal_id

Crossref journal ID

cr_year

Earliest publication year (Crossref field `issued`)

cc

Normalized Creative Commons variant. Ordered factor by license variant permissiveness

cc_total

Number of articles under Creative Commons variant

jn_all

Yearly journal output

prop

Proportion of CC licensed articles

Examples

cc_jn_ind
#> # A tibble: 43,752 × 6
#>    cr_journal_id cr_year cc          cc_total jn_all    prop
#>            <int> <fct>   <fct>          <int>  <int>   <dbl>
#>  1        144347 2019    NA                 0    256 0      
#>  2         98185 2021    CC BY              1    256 0.00391
#>  3         98185 2021    CC BY-NC-ND        1    256 0.00391
#>  4        208685 2021    CC BY-NC-ND        1    256 0.00391
#>  5         24919 2019    CC BY-NC-ND        1    256 0.00391
#>  6          3428 2018    CC BY-SA           1    256 0.00391
#>  7          3428 2018    CC BY-NC-ND        3    256 0.0117 
#>  8            16 2018    CC BY-NC           3    256 0.0117 
#>  9          5668 2022    CC BY              3    256 0.0117 
#> 10          3428 2018    CC BY-NC-SA        3    256 0.0117 
#> # … with 43,742 more rows