First author country affiliations per journal, year and Creative Commons license

This dataset contains the number and proportion of open access articles in hybrid journals by country, year and Creative Commons license variant.

Usage

cc_openalex_inst_jn_ind

Format

An object of class tbl_df (inherits from tbl, data.frame) with 664657 rows and 6 columns.

Details

Country affiliations were determined using the first author affiliation as listed by OpenAlex. A first author is often considered as lead author(https://en.wikipedia.org/wiki/Lead_author) who has usually undertaken most of the research presented in the article, although author roles can vary across disciplines.

Note that full counting in which every original article or review was counted once per country of affiliation of the first authors was applied. Because first authors can have multiple affiliations from different countries, don't use this dataset, but cc_jn_ind to determine a journal's publication volume.

Variables:

cr_journal_id

Crossref journal ID

cr_year

Earliest publication year (Crossref field `issued`)

country_code

The country where this institution is located, represented as an ISO two-letter country code. (OpenAlex field `country_code`)

cc

Normalized Creative Commons variant. NA represents articles, which were not provided under a CC license

articles_under_cc_variant

Number of articles under Creative Commons variant, grouped by journal, year and country affiliation

articles_total

Yearly journal output by year and country affiliation

Examples

# Scientometrics (Crossref journal ID: 2795)
  subset(cc_openalex_inst_jn_ind,
    cr_journal_id %in% "2795" & country_code %in% c("DE", "CN"))
#> # A tibble: 20 × 6
#>    cr_journal_id cr_year country_code cc    articles_under_cc_va… articles_total
#>            <int>   <int> <chr>        <chr>                 <int>          <int>
#>  1          2795    2022 CN           NA                       31             31
#>  2          2795    2022 DE           CC BY                     8              8
#>  3          2795    2021 CN           NA                       86             92
#>  4          2795    2021 CN           CC BY                     6             92
#>  5          2795    2021 DE           CC BY                    26             29
#>  6          2795    2021 DE           NA                        3             29
#>  7          2795    2020 CN           NA                       86             94
#>  8          2795    2020 DE           CC BY                    30             35
#>  9          2795    2020 DE           NA                        5             35
#> 10          2795    2020 CN           CC BY                     8             94
#> 11          2795    2019 CN           NA                       66             67
#> 12          2795    2019 DE           NA                       12             18
#> 13          2795    2019 DE           CC BY                     6             18
#> 14          2795    2019 CN           CC BY                     1             67
#> 15          2795    2018 CN           NA                       65             65
#> 16          2795    2018 DE           NA                       19             24
#> 17          2795    2018 DE           CC BY                     5             24
#> 18          2795    2017 CN           NA                       71             71
#> 19          2795    2017 DE           NA                       21             28
#> 20          2795    2017 DE           CC BY                     7             28