Prevalence of Creative Commons licenses by variant, year and journal

This dataset contains the number and proportion of open access articles with Creative Commons license (CC) by license variant and year for hybrid journals included in the cOAlition S Journal Checker Tool.

Usage

jn_ind

Format

An object of class tbl_df (inherits from tbl, data.frame) with 94496 rows and 6 columns.

Details

Publication period is 2017 - 2022.

Journal's article volume was calculated using Crossref metadata snapshot. Note that only articles published in regular issues aside from supplements containing conference contributions like meeting abstracts, indicated by non-numeric pagination, were included. Also, non-scholarly journal content, such as the table of contents were excluded. In doing so, we followed Unpaywall's paratext recognition approach, which we expanded to include patterns indicating corrections.

CC licenses were also identified through Crossref. License information for author accepted manuscripts ("aam") were not considered.

Variables:

issn_l

Linking ISSN

cr_year

Earliest publication year (Crossref field `issued`)

cc

Normalized Creative Commons variant. Ordered factor by license variant permissiveness

cc_total

Number of articles under Creative Commons variant

jn_all

Yearly journal output

prop

Proportion of CC licensed articles

Examples

jn_ind
#> # A tibble: 94,378 × 6
#>    issn_l    cr_year cc          cc_total jn_all    prop
#>    <chr>     <fct>   <fct>          <int>  <int>   <dbl>
#>  1 0001-0782 2017    NA                 0    281 0      
#>  2 0001-0782 2018    NA                 0    302 0      
#>  3 0001-0782 2019    NA                 0    280 0      
#>  4 0001-0782 2020    NA                 0    268 0      
#>  5 0001-0782 2021    NA                 0    318 0      
#>  6 0001-0782 2022    NA                 0    242 0      
#>  7 0001-1541 2017    CC BY-NC           4    454 0.00881
#>  8 0001-1541 2017    CC BY              4    454 0.00881
#>  9 0001-1541 2018    CC BY-NC-ND        2    342 0.00585
#> 10 0001-1541 2018    CC BY              7    342 0.0205 
#> # … with 94,368 more rows